Refactor: extract structs
