+++ gogorillarest/pkg/Data.go
@@ -1,0 +1,8 @@
+package pkg
+
+type Data struct {
+	PlayerInitialLives      string `yaml:"player_initial_lives"`
+	UiPropertiesFileName    string `yaml:"ui_properties_file_name"`
+	GameProperties          string `yaml:"game.properties"`
+	UserInterfaceProperties string `yaml:"user-interface.properties"`
+}
