+++ gogorillarest/pkg/MetaData.go
@@ -1,0 +1,5 @@
+package pkg
+
+type Metadata struct {
+	Name string `yaml:"name"`
+}
